feat(dsp): rotating chords accept just ratios, not only semitones
In a microtonal synth the chord intervals should be able to be exact
just intervals, so a chord-sequence token can now be either decimal
semitones ("7", "-5", "7.02") or a just ratio ("3:2", "5/4"). Ratios
convert to fractional semitones (12*log2(a/b)), so a 3:2 fifth is the
exact 701.96c rather than a rounded 700c.

To preserve that precision the interval storage changes from int8_t
semitones to float semitones end to end (chordSeq, chordSeqFlat, the
voice read). The raw text is what persists in state, so "3:2" round-trips
as typed; only the parsed array is fractional. The application path was
already float (pow(2, semis/12)), so the audio side needed no change.

Parsing lives in a static, unit-tested VaneProcessor::parseChordInterval
that rejects junk as NaN (so a stray token is skipped, not silently
turned into a 0-semitone unison). The matrix UI editor accepts the new
tokens and documents both forms.

Tests: ChordTests covers decimal/ratio/junk; 3:2 asserted distinct from
a rounded 700c. 44 suites / 70215 assertions pass; all formats build
clean; startup gate exit 142, 0 assertions.

+++ Source/Synth/SynthVoice.cpp
@@ -458,7 +458,7 @@ void SynthVoice::noteStarted()
         for (int j = 0; j < kMaxUnison - 1; ++j) {
             const int len = chordLens[j];
             chordInterval[j] = (len > 0)
-                ? static_cast<float>(chordSeqFlat[j * kChordSteps + (idx % len)])
+                ? chordSeqFlat[j * kChordSteps + (idx % len)]
                 : 0.0f;
         }
     }
